+++ changelog.md
@@ -1,5 +1,5 @@
 ## Upcoming
-* Added back Mobs feature: Equipment, Misc Mobs and Zombie Siege
+* Added back Mobs feature: Equipment, Misc Mobs, Spawning and Zombie Siege
 
 ## 5.1.0.0-alpha
 * Added back Absorption Armor, Armor rework, regen absorption, Bows, Snowballs, Piercing Damage, Unfair oneshot and Misc Stats